requirements/production.txt
-r base.txt
kb-base==0.3.24
kb-block==0.3.26
kb-chat==0.0.10
kb-compose==0.1.23
kb-contact==0.0.68
kb-crm==0.1.76
kb-enquiry==0.1.45
kb-finance==0.1.17
kb-gallery==0.0.10
kb-gdpr==0.0.31
kb-googl==0.0.05
kb-googl-crm==0.0.04
kb-invoice==0.2.08
kb-login==0.2.30
kb-mail==0.1.85
kb-report==0.0.65
kb-search==0.1.39
kb-stock==0.1.38

